+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed: `selector-class-pattern` now ignores fractional keyframes selectors.
 - Fixed: `selector-max-specificity` now ignores selectors containing the `matches()` pseudo-class, and warns if the underlying `specificity` module cannot parse the selector.
 - Fixed: `selector-no-type` with secondary option `ignore: ["descendant"]` will now resolve nested selectors.
+- Added: `custom-property-no-empty-line-before` rule.
 
 # 7.0.2